feat: add event types and update event request validation
Introduce an `EventType` enum to categorize events as in-person, online, 
or hybrid. Update the `ParticipationType` enum to include paid and free 
options. Modify the `EventController` to authorize organizers and 
refactor the `UpdateEventRequest` to change 'address' to 'location', 
make 'feature_image' optional, and allow nullable 'end_date'. Add a 
migration to make 'end_date' and 'registration_deadline' nullable in 
the database.